Top down approach
First attempt. Did not follow TDD rules. While trying to write the first test, I knew it wasn't what I wanted.
At the end, what I did was to write the loop in MainApp.java first and then the method I wanted to call with FizzBuzz.java.
It was a top down approach.